Fix remaining Windows platform issues
- Kernel.File.Open.swift: Cast FILE_SHARE_* constants to DWORD
- Kernel.Mmap.swift: Fix Bool comparisons for UnmapViewOfFile, FlushViewOfFile, VirtualProtect

+++ Sources/Kernel/Kernel.File.Open.swift
@@ -283,7 +283,7 @@ extension Kernel.File.Open.Options {
         /// This is a documented stability guarantee matching common POSIX expectations.
         @usableFromInline
         internal static var windowsShareMode: DWORD {
-            return FILE_SHARE_READ | FILE_SHARE_WRITE | FILE_SHARE_DELETE
+            return DWORD(FILE_SHARE_READ) | DWORD(FILE_SHARE_WRITE) | DWORD(FILE_SHARE_DELETE)
         }
     }
 #endif
